fix: move EditNFTDescriptionModal.tsx up one folder, and remove unnecessary index file
