--- src/components/[guild]/RoleCard/components/EditNFTDescription/index.ts
@@ -1,3 +1,0 @@
-import { EditNFTDescriptionModal } from "./EditNFTDescriptionModal"
-
-export default EditNFTDescriptionModal
