+++ src/pages/[guild]/collect/[chain]/[address].tsx
@@ -25,7 +25,6 @@ import useNftDetails from "components/[guild]/collect/hooks/useNftDetails"
 import useGuild from "components/[guild]/hooks/useGuild"
 import useGuildPermission from "components/[guild]/hooks/useGuildPermission"
 import ReportGuildButton from "components/[guild]/ReportGuildButton"
-import EditNFTDescriptionModal from "components/[guild]/RoleCard/components/EditNFTDescription"
 import { ThemeProvider, useThemeContext } from "components/[guild]/ThemeContext"
 import { usePostHogContext } from "components/_app/PostHogProvider"
 import CardMotionWrapper from "components/common/CardMotionWrapper"
@@ -51,6 +50,7 @@ import { SWRConfig } from "swr"
 import { Guild } from "types"
 import fetcher from "utils/fetcher"
 import { PencilSimple } from "phosphor-react"
+import { EditNFTDescriptionModal } from "components/[guild]/RoleCard/components/EditNFTDescriptionModal"
 
 type Props = {
   chain: Chain
