+++ src/platforms/ContractCall/ContractCallCardMenu.tsx
@@ -1,10 +1,10 @@
 import EditRewardAvailabilityMenuItem from "components/[guild]/AccessHub/components/EditRewardAvailabilityMenuItem"
 import RemovePlatformMenuItem from "components/[guild]/AccessHub/components/RemovePlatformMenuItem"
 import PlatformCardMenu from "../../components/[guild]/RolePlatforms/components/PlatformCard/components/PlatformCardMenu"
-import EditRewardDescriptionModal from "../../components/[guild]/RoleCard/components/EditNFTDescription"
 import { MenuItem, useDisclosure } from "@chakra-ui/react"
 import useGuild from "../../components/[guild]/hooks/useGuild"
 import { PencilSimple } from "phosphor-react"
+import { EditNFTDescriptionModal } from "../../components/[guild]/RoleCard/components/EditNFTDescriptionModal"
 
 type Props = {
   platformGuildId: string
@@ -28,7 +28,7 @@ const ContractCallCardMenu = ({ platformGuildId }: Props): JSX.Element => {
         <RemovePlatformMenuItem platformGuildId={platformGuildId} />
       </PlatformCardMenu>
 
-      <EditRewardDescriptionModal
+      <EditNFTDescriptionModal
         isOpen={isOpen}
         onClose={onClose}
         guildPlatform={guildPlatform}
